extra compiler warnings, more type correctness, support for tree or linear process spawning, added option to print run configuration at program start

+++ arraywalk/csv.h
@@ -5,5 +5,5 @@
 #include <stdint.h>
 #include <stdio.h>
 
-typedef uint_least64_t nsec_t;
+typedef long long nsec_t;
 void CSV_LogTimings(FILE * log, struct walkArray * wa, nsec_t nsec, nsec_t stddev);
